requirements.txt
python-telegram-bot==21.6
APScheduler==3.10.4
requests==2.32.3